+++ src/__init__.py
@@ -1,0 +1,1 @@
+# The "src" folder includes the main code of the password strength checker.