+++ CHANGELOG.md
@@ -1,6 +1,10 @@
 # CHANGELOG:
 
-## V1.1:
+## v1.1.1:
+- Fixed a bug that allowed passwords with too many repeated characters to be determined "strong"
+- Fixed a bug that caused the dictioanry attack vulnerability checker to look for a variable that I forgot to put causing an error and ending the program
+
+## v1.1:
 - Added CHANGELOG.md
 - Organized the files further
 - Fixed a bug that made the dictionary attack vulnerability detector less strict than it should be